silenced visual studio windows warnings

+++ common.c
@@ -9,7 +9,7 @@ float
 percentage(unsigned long int n, unsigned long int d)
 {
   /* TBD: Error check here */
-  float percent = ((float)n / (float)d)*100.0;
+  float percent = (float)(n / d* 100.0);
   return percent;
 }
